refactor: bindads token code
