+++ router/batchrouter/asyncdestinationmanager/bing-ads/common/token.go
@@ -1,4 +1,4 @@
-package bingads_token
+package common
 
 import (
 	"encoding/json"
